refactor: split Copilot provider internals
